Rename xml_validator
