+++ easy2homeassistant/xml_validator.py
